R3-12: mheard lat/lon to float, with the precision cost measured
mheardLat[]/mheardLon[] were double. Measured per env, matching the audit:
-648 B on twelve envs, -640 on ten, -400 on one, -240 on nine, 0 on the two
safeboot envs which have no mheard. Flash +608 B across all 34 (~19 B each) for
the conversions at the assignment sites.

PRECISION MEASURED, NOT ARGUED. 600 000 random coordinate pairs through a float
round trip: worst absolute error 0.00000763 deg, about 0.85 m. At the web GUI's
%06.3lf/%07.3lf, 0.15 % of coordinates render a different last digit -- roughly
one in 670, and only where the true value sits on a rounding boundary. The
mheard list is a neighbourhood display, not navigation, so 0.85 m is not a
concern; the last-digit render is worth writing down so that a future capture
diff showing one is recognised as this and not as a regression.

Two things this row was not, until they were looked at:

It is not a single-file change. Both arrays are re-declared by hand as `extern`
in src/lora_functions.cpp and src/web_functions/web_functions.cpp. Missing one
links SILENTLY and reads the array at the wrong width -- no compiler sees both
TUs, and the linker matches names, not types. All three spellings changed
together, and carve_extern_lint.py now covers those files (previous commit).

It is not RAM-only. Both arrays are written raw into /mheard.dat on the T-Deck
(file.write((uint8_t*)mheardLat, sizeof(mheardLat))), so the file layout
changes with the type. That is already handled: the load path compares
file.size() against the sum of the sizeof()s and deletes the file on mismatch,
so an existing file is lost once and rebuilt. Self-healing, but it happens, and
it is now written at the declaration instead of being rediscovered.

Gate: 34/34 board envs, 31/31 native envs, test/golden/selftest.sh exit 0,
carve_extern_lint clean over 52 externs.

+++ src/lora_functions.cpp
@@ -112,8 +112,8 @@ void loraDeepSleep()
 extern unsigned long iReceiveTimeOutTime;
 
 extern char mheardCalls[MAX_MHEARD][10]; //Ringbuffer for MHeard Key = Call
-extern double mheardLat[MAX_MHEARD];
-extern double mheardLon[MAX_MHEARD];
+extern float mheardLat[MAX_MHEARD];   // R3-12: war double
+extern float mheardLon[MAX_MHEARD];   // R3-12: war double
 extern int mheardAlt[MAX_MHEARD];
 
 #include "TinyGPSPlus.h"
